+++ CHANGELOG.md
@@ -5,7 +5,9 @@ ember-django-adapter Changelog
 Master
 ------
 * [ENHANCEMENT] Add support for HyperlinkedRelatedFields.
-
+* [INTERNAL] Updated ember-cli version to latest (1.13.1)
+* [ENHANCEMENT] Adapter now supports ember-data 1.13.5
+  ([#96](https://github.com/dustinfarris/ember-django-adapter/pull/108))
 
 0.5.6
 -----